[MCUX-78748] Fix coverity issues of components_sdu
[Description]
SDU_CreateHandle(): Cast from 8 bit width expression "8 | (config->data_tx_format << 12) | (config->data_rd_format << 14)" to a wider 32 bit type.

[Root Cause]
Source code against rules of MISRA.

[Fix]
Cast from 8 bit width value to a 32 bit width before used as 32 bit width in SDU_CreateHandle().

[Testing]
Local coverity scan ok, ncp_host/ncp_device connect to AP and ping ok.

+++ sdu/fsl_adapter_sdu.c
@@ -439,8 +439,8 @@ static sdioslv_handle_t SDU_CreateHandle(sdioslv_handle_config_t *config)
     sdu_fsr->CardIntMode   = 0;//0x00000C02;//CARD_INT_MODE_MSK;
     sdu_fsr->CmdPortConfig =
         (uint32_t)((config->cmd_tx_format << (uint8_t)CMD_TX_LEN_BIT_OFFSET) | (config->cmd_rd_format << (uint8_t)CMD_RD_LEN_BIT_OFFSET));
-    sdu_fsr->Config2 = (uint32_t)((uint8_t)CONFIG2_DEFAULT_SETTING | (config->data_tx_format << (uint8_t)(CONFIG2_TX_LEN_BIT_OFFSET)) |
-        (config->data_rd_format << (uint8_t)(CONFIG2_RD_LEN_BIT_OFFSET)));
+    sdu_fsr->Config2 = (uint32_t)CONFIG2_DEFAULT_SETTING | (((uint32_t)config->data_tx_format) << ((uint8_t)CONFIG2_TX_LEN_BIT_OFFSET)) |
+        (((uint32_t)config->data_rd_format) << ((uint8_t)CONFIG2_RD_LEN_BIT_OFFSET));
 
     for (i = 0; i < SDU_PORT_MAX_TRANSFER; i++)
     {
